fix(managed,mcp): аллокатор адреса пира — leaf-пакет peerip ради сборки mcp-dev на darwin
CI (job mcp-portable) кросс-собирает cmd/mcp-dev под darwin/arm64, а
после переноса аллокатора в internal/managed фейк mcptest потянул
osdetect с syscall.Uname и сборка упала. Правило по-прежнему одно:
peerip.NextFree применяет managed.AddPeer при пустом TunnelIP, адаптер
MCP ничего не выделяет, фейк импортирует только leaf-пакет. README
внутри internal/mcp фиксирует это ограничение со ссылкой на job.
