+++ CONTRIBUTING.md
@@ -121,7 +121,7 @@ cd frontend && VITE_API_TARGET=http://127.0.0.1:8081 npm run dev:mock
 
 ### MCP в mock-режиме
 
-`npm run dev:mock:proxy` дополнительно поднимает `go run ../cmd/mcp-dev` на `http://127.0.0.1:8090/mcp` (нужен Go на Linux — `mcptest` тянет `internal/managed`; отключить — `MOCK_MCP=0`). Подключение: `claude mcp add --transport http awgm-dev http://127.0.0.1:8090/mcp`. Данные — фейковые (`internal/mcp/mcptest`), карточка ключей в UI работает через `mock-proxy.mjs` и с dev-сервером не связана.
+`npm run dev:mock:proxy` дополнительно поднимает `go run ../cmd/mcp-dev` на `http://127.0.0.1:8090/mcp` (нужен Go; отключить — `MOCK_MCP=0`). Подключение: `claude mcp add --transport http awgm-dev http://127.0.0.1:8090/mcp`. Данные — фейковые (`internal/mcp/mcptest`), карточка ключей в UI работает через `mock-proxy.mjs` и с dev-сервером не связана.
 
 ## Процесс работы над фичей
